Rework

+++ src/asmutils.rs
@@ -13,9 +13,6 @@ pub fn disassemble<'a, D>(
     size: usize,
 ) -> OwnedInsn<'a> {
     let mut mem = [0u8; 8];
-    emu.mem_read(address, &mut mem[..size as usize]).unwrap();
-    (&capstone
-        .disasm_count(&mem[..size as usize], address, 1)
-        .unwrap()[0])
-        .into()
+    emu.mem_read(address, &mut mem[..size]).unwrap();
+    (&capstone.disasm_count(&mem[..size], address, 1).unwrap()[0]).into()
 }
